refactor: 미니플레이어 컨트롤러 분리
